Calibrated SHMS DCs.
The calibration scripts work in a very specific way. First you have to change the
`wire_drift_times.h` to read the correct run. Then you have to change the
`pdc_plane_time_zero` in the parameter file to `1280`, since it is hardcoded.
Then you replay the run with these values and run the calibration. Then you
get the correct corrections to the time zero.