Target masses where the value was zero were validated (gtargmass_amu = 0.0) as being Home, Empty, or ??. One run, 4435, was ??.  This corresponded to target still moving (Junk run based on runsheets). Most were home: 5036 - 5037, 5209, 5314, 5577 - 5578, 6076, 6092 - 6093, 6139 - 6140. Some runs were incorrectly 0.0 when they should have been He3: 5789 - 5791 <-- Changed to He3 based on encoder position.
The target was not listed for runs 6474 - 6778 because of a change in encoder position.  This was corrected by looking at the logbook entry title for that range.  A short temporary file was created and filled within the ranges.  Two runs did not have logbook entries:
6662 <-- At 10 Dummy based on evio2xml -max 100 | grep "BDS".
6675 <-- At Ca-40 based on evio2xml -max 100 | grep "BDS".
This makes sense based on runs 6663 and 6676.