COINCIDENCE replay updates
Lots of changes for coincidence replays. Most grabbed from Abhyuday. Small changes in COIN standard.database to get correct optics based on SHMS momentum.

Small change to hmsflags_fa22.param. Line with equation for hms momentum offset is commented out and then set to zero. The equation was used for HMS pass 1. COIN would like to use 0 for now. Probably should make a second file for this so that toes are stepped on. I'll open an issue to remind myself to do that.

Co-Authored-By: Abhyuday Sharda <117782573+ab6361@users.noreply.github.com>