New BCM current maps
Updated bcm current maps so that currents are no longer rounded to an integer. This push includes all HMS, SHMS, and COIN runs.

Note that the map for run 19470 is deleted in this request and not replaced. It is a junk run that was previously replayed and no longer is included in replay jobs.